+++ src/report/report-properties.ts
@@ -1,4 +1,4 @@
 export interface ReportProperties {
   maxSize?: number
-  errorOnSeverities: string[]
+  considerErrorOnSeverities: string[]
 }